+++ evaluating-probes/src/probes/sae_probe.py
@@ -36,7 +36,8 @@ class SAEProbeNet(nn.Module):
         self.sae_feature_dim = sae_feature_dim
         self.aggregation = aggregation
         self.device = device
-        self.linear = nn.Linear(sae_feature_dim, 1).to(self.device)
+        # Create linear layer with float32 dtype for mixed precision training
+        self.linear = nn.Linear(sae_feature_dim, 1, dtype=torch.float32).to(self.device)
 
     def forward(self, x: torch.Tensor, mask: torch.Tensor) -> torch.Tensor:
         # x: (batch, seq, sae_feature_dim), mask: (batch, seq)
@@ -172,7 +173,7 @@ class SAEProbe(BaseProbe):
         for i in tqdm(range(0, len(flattened), self.encoding_batch_size), desc="Encoding activations"):
             batch = flattened[i:i+self.encoding_batch_size]
             # Convert batch to tensor and move to device only for this batch
-            batch_tensor = torch.tensor(batch, dtype=torch.float32, device=self.device)
+            batch_tensor = torch.tensor(batch, dtype=torch.float16, device=self.device)
             encoded_batch = sae.encode(batch_tensor).cpu().detach().numpy()
             encoded_list.append(encoded_batch)
